Version 0.38b --------------------------------------     + Fixed the z-coordinate Inversion when exporting/importing a Mesh     + Added a total Face/Vertex Count     + Vertex count will now Display the Number of really used Vertices (not the Number of referenced Vertices)

+++ fullsimpe/SimPe RCOL/ImageLoader.cs
@@ -564,7 +564,7 @@ namespace SimPe.Plugin
 				}
 
 				abits = abits << 3;
-				abits = abits | (long)index;
+				abits = abits | (int)index;
 			}
 
 			ushort abits1 = (ushort)((abits >> 32) & 0x0000ffff);
